Add CircleCI config
This is a bit of a hacky mess, mostly due to the fact that there don't
seem to be working CircleCI images for Ruby 1.8.7. I'll continue to
revisit down the road and work on cleaning it up. It's also going to be
very slow because it compiles/installs Ruby from scratch each build. For
now this is tolerable, since there is nobody but myself actively working
on this project, and I'm not in a huge rush.

**Notes**

* Next step is probably introducing a `Dockerfile` and slowly moving
some of this stuff to a custom image. I'm leaving some commented stuff
for now for future reference.
* I copied the same `source` command for `rvm` into each `run` block. It
looks like it's not getting loaded automatically, so we need to handle
it. Probably a good future step would be to figure out which `*rc` files
get loaded and put it in there.
* I run the test folders one-by-one because the tests in
`spec/integration/` aren't working at the moment. I did a quick check,
but it doesn't appear that older versions of RSpec have the ability to
filter out certain folders. It might be good down the road to split up
each folder into a different job, but it'll probably slow things down in
the short term, since we're re-installing Ruby for each one.