add specs for IssueInvitationsController (#104)
As near as I can tell, this controller is not used from anywhere in the
application. It looks like it just got added in during a JQuery upgrade
at one point. For now I've added tests, but down the road once we have
better feature testing, we may be able to remove it if we don't find any
usage.

I removed the `format.html` block from the `new` action because there is
no corresponding view file, so it will always throw an error if hit.